# documentacion/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger Javascript Partida API",
    "description": "Javascript Partida API documentation"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "get": {
        "description": "Returns all users.\n",
        "parameters": [],
        "operationId": "findUsers",
        "responses": {
          "200": {
            "description": "Return all users from the application",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            },
            "500": {
              "description": "General error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "post": {
        "description": "Creates a new user in the application.",
        "operationId": "addUser",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User to add to the application",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return new user.",
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          },
          "500": {
            "description": "General error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Returns a user based on a single ID",
        "operationId": "find user by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to fetch",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Return user found.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "General error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "patch": {
        "description": "Update a user in the application.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "User to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return updated user.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "General error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "description": "deletes a single user based on the ID supplied",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns user deleted",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "General error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Mongo Object Id"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "description": "Email format"
        },
        "password": {
          "type": "string"
        },
        "color":  {
          "type": "string"
        },
        "image": {
          "type": "string",
          "description": "Image in base64"
        }
      }
    },
    "NewUser": {
      "type": "object",
      "required": [
        "name",
        "email",
        "password",
        "color"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "description": "Valid email format"
        },
        "password": {
          "type": "string"
        },
        "color":  {
          "type": "string",
          "description": "Hexadecimal color"
        },
        "image": {
          "type": "string",
          "description": "Base64 image"
        }
      }
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}